Add edit view to deck
