+++ client/src/pages/DeckDetails.jsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from "react";
 import {
   ButtonGroup,
+  Fab,
   FormControl,
   Grid2 as Grid,
   InputLabel,
@@ -11,7 +12,7 @@ import {
 } from "@mui/material";
 import { BasePage } from "./BasePage";
 import styled from "styled-components";
-import { useParams } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 import { getDeck } from "../api/decks";
 import { useCardList } from "../contexts/CardContext";
 import { Loading } from "../components/Loading";
@@ -20,6 +21,8 @@ import { AspectIcons } from "../components/AspectIcons";
 import { CardDialog } from "../components/CardDialog";
 import { SWUCardDeck } from "../components/SWUCardDeck";
 import { sortList } from "../utils/sortCardList";
+import { useAuth } from "../contexts/AuthContext";
+import EditNoteIcon from "@mui/icons-material/EditNote";
 
 const Styles = {
   ButtonGroup: styled(ButtonGroup)`
@@ -44,11 +47,21 @@ const Styles = {
     margin-top: 10px;
     flex-direction: column;
   `,
+  EditButton: styled(Fab)`
+    position: fixed;
+    bottom: 50px;
+    right: 0;
+    width: 70px;
+    border-radius: 25px 0 0 25px;
+    z-index: 50;
+  `,
 };
 
 export function DeckDetails() {
   const { deckId } = useParams();
   const { getCardData, cardList } = useCardList();
+  const { account } = useAuth();
+  const navigate = useNavigate();
   const [deckData, setDeckData] = useState({});
   const [leader, setLeader] = useState({});
   const [base, setBase] = useState({});
@@ -126,6 +139,14 @@ export function DeckDetails() {
 
   return (
     <BasePage>
+      {account?._id === deckData?.ownerId ? (
+        <Styles.EditButton
+          color="success"
+          onClick={() => navigate("/deck/edit/" + deckId)}
+        >
+          <EditNoteIcon />
+        </Styles.EditButton>
+      ) : null}
       <CardDialog
         cardData={clickedCard}
         handleCloseDialog={handleCloseDialog}
